requirements.txt
# DevOps tools and libraries
black==22.3.0
click==8.1.3
pytest==7.1.3
pytest-cov==4.0.0
requests==2.26.0

# Rust-based linter
ruff==0.0.284

# For environment variable management
python-dotenv==0.21.1


databricks-sql-connector==2.7.0